refactor: replace `hooks.abstract_hook` imports with `shared.hook` and rename `Policy` to `AbstractPolicy` across the codebase
