+++ python/hedera_agent_kit/policies/max_recipients_policy.py
@@ -11,8 +11,8 @@ from hedera_agent_kit.shared.parameter_schemas import (
     TransferNonFungibleTokenWithAllowanceParametersNormalised,
     TransferNonFungibleTokenParametersNormalised,
 )
-from hedera_agent_kit.shared.policy import Policy
-from hedera_agent_kit.hooks.abstract_hook import PostParamsNormalizationParams
+from hedera_agent_kit.shared.policy import AbstractPolicy
+from hedera_agent_kit.shared.hook import PostParamsNormalizationParams
 from hedera_agent_kit.plugins.core_account_plugin import core_account_plugin_tool_names
 from hedera_agent_kit.plugins.core_token_plugin import core_token_plugin_tool_names
 
@@ -86,7 +86,7 @@ _BUILTIN_STRATEGIES: Dict[str, Callable[[Any], int]] = {
 }
 
 
-class MaxRecipientsPolicy(Policy):
+class MaxRecipientsPolicy(AbstractPolicy):
     """
     Limits the maximum number of recipients allowed in a single transfer / airdrop call.
