+++ python/hedera_agent_kit/shared/hook.py
