+++ python/test/unit/hook/hcs_audit_trail_hook_test.py
@@ -4,7 +4,7 @@ from unittest.mock import MagicMock, patch
 
 from hedera_agent_kit.hooks.hcs_audit_trail_hook import HcsAuditTrailHook
 from hedera_agent_kit.hooks.utils import stringify_recursive
-from hedera_agent_kit.hooks.abstract_hook import (
+from hedera_agent_kit.shared.hook import (
     PreToolExecutionParams,
     PostSecondaryActionParams,
 )
